SQL, projection, GET games by list

+++ src/main/java/com/devsuperior/dslist/entities/Game.java
@@ -10,7 +10,7 @@ public class Game {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private Long id;
     private String title;
 
     @Column(name = "game_year")
@@ -29,7 +29,7 @@ public class Game {
     public Game() {
     }
 
-    public Game(int id, String title, Integer year, String genre, String platforms, Double score, String imgUrl, String shortDescription, String longDescription) {
+    public Game(Long id, String title, Integer year, String genre, String platforms, Double score, String imgUrl, String shortDescription, String longDescription) {
         this.id = id;
         this.title = title;
         this.year = year;
@@ -41,11 +41,11 @@ public class Game {
         this.longDescription = longDescription;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
